Create script inside index.html file
Created a simple script for one round of the Rock, Paper, Scissors game.
The script includes the function getComputerChoice() to generate the
game options at random by making use of the function Math.random(), and
the playRound() function to compare the player's choice and the
computer's choice to decide who won the round. The original function to
normalize the format of the player's choice was commited out for
testing.