Fix looping bugs for several functions, feat Round()
Instead of using a recursive soluntion involving the if condition as
tried before, used the iterative approach to loop the prompt until the
player chooses one of the correct options. Also rearranged the position
of the functions so that they ran in the correct order and to get the
game() function to play five rounds. Before the rearrangement, the code
would not recognize some variables. Included a Round function to track
the number of rounds.