Combine functions under one loop
Originally I tried to separate functions for retrieving the player's
selection, computer's selection, and the single-round game result.
Although this allowed the correct player and computer selections, it
did not return the correct comparisons to determine the winner. I kept
the computer function separate, but combined player selection and game
result under the same forEach loop.