Orangize functions within forEach loop
Moved the function getComputerChoice inside the forEach loop to make the
code more organized. Also moved the forEach loop inside a for loop such
that (let i = 1; i <= 5; i++), which is supppose to run the game for
five rounds. Instead, after one of the playable buttons was clicke, the
game ran through 5 rounds with the same result. Another issue: the score
is not properly recorded.