Fix scoring issues and make style changes
Removed the function countScore() and instead included playerScore++ and
computerScore++ to the function roundWinner() such that a point is added
to the correct variable when either player or computer wins a point.
Able to display the round number after declaring new numberRound
variable and including numberRound++ to the for Each loop
addEventListerner. Also rearranged HTML elements and updated CSS to make
the webpage more presentable. Now trying to debug code that is suppose
to declare game winner after either opponent wins 5 points.