Fix function winner() and change a variable
Function winner() now stops the game when either opponent reaches 5
points. I had to reorganized the order of functions inside the forEach
loop. I renamed the function startAgain() as startRound() and the
varialbe playAgain as nextRound for clarity and so that I can later
change the code to have an option to play the game again.