Fix player and computer style, and move playAgain
The elements displaying the player and computer selections were
displaying the wrong style when playerSelection was the same as
computerSelection. Creating a new function roundStyle() to define the
style based on the result of roundWinner() corrected this issue. Also
removed const playAgain = document.createElement('button') from the
function winner() and instead created a button with id="playAgain" and
included playAgain.remove() about the forEach loop to make the button's
function independent of the loop.