Fix "play again" button
Renamed the function playAgain() as playGameAgain() for clarity and
moved it after the forEach loop within the function gameLoop(). To get
the rock, paper, scissors buttons to appear and work, needed to include
article.appendChild(choices) as well as choices.appendChild(rock),
choices.appendChild(paper), and choices.appendChild(scissors). Although
the game continues to play, the score does not appear.